added get user fav recipe, delete user diet , delete user endpoint and minor changes to other routes

+++ Backend/routes/maria/user.js
@@ -10,7 +10,7 @@ router.get('/:id', (req, res) => {
     db.query(sql, [userId])
         .then((rows) => {
             if (rows.length === 0) {
-                res.status(404).json({ message: 'User not found' });
+                res.status(400).json({ message: 'User not found' });
             } else {
                 rows[0].userId = Number(rows[0].userId);
                 res.json(rows[0]);
@@ -24,6 +24,10 @@ router.get('/:id', (req, res) => {
 router.put('/:id', (req, res) => {
     const userId = req.params.id;
     const { username, password, image } = req.body;
+
+    if (!username || !password) {
+        return res.status(400).json({ message: 'Please provide username and password' });
+    }
 
     const saltRounds = 10;
     const passwordHash = bcrypt.hashSync(password, saltRounds);
@@ -42,27 +46,57 @@ router.put('/:id', (req, res) => {
         });
 });
 
+router.delete('/:id', (req, res) => {
+    const userId = req.params.id;
+
+    const sql = 'DELETE FROM user WHERE userId = ?';
+    db.query(sql, [userId])
+        .then((result) => {
+            if (result.affectedRows === 0) {
+                res.status(400).json({ message: 'User not found' });
+            } else {
+                res.json({ message: 'User deleted successfully' });
+            }
+        })
+        .catch((err) => {
+            res.status(500).json({ error: err.message });
+        });
+});
+
 router.post('/:id/diet', async (req, res) => {
     const userId = req.params.id;
-    const { diet } = req.body;
+    const diets = req.body.diets;
+
+    if (!diets) {
+        return res.status(400).json({ message: 'Please provide diets' });
+    }
+
+    const dietArr = diets.split(',').map(item => item.trim());
 
     try {
+        const userExistsSQL = 'SELECT 1 FROM user WHERE userId = ?';
+        const userResult = await db.query(userExistsSQL, [userId]);
+
+        if (userResult.length === 0) {
+            return res.status(404).json({ message: 'User not found' });
+        }
+
         const deleteResult = await new Promise((resolve, reject) => {
             const sqlDelete = 'DELETE FROM userdiet WHERE userId = ?';
             db.query(sqlDelete, [userId])
-                .then((result) => resolve(result))
+                .then(result => resolve(result))
                 .catch(reject);
         });
 
         let successfulInserted = 0;
 
-        for (let i = 0; i < diet.length; i++) {
-            const dietName = diet[i];
+        for (let i = 0; i < dietArr.length; i++) {
+            const dietName = dietArr[i];
 
             const insertResult = await new Promise((resolve, reject) => {
                 const sqlInsert = 'INSERT INTO userdiet (userId, diet) VALUES (?,?)';
                 db.query(sqlInsert, [userId, dietName])
-                    .then((result) => resolve(result))
+                    .then(result => resolve(result))
                     .catch(reject);
             });
 
@@ -71,7 +105,7 @@ router.post('/:id/diet', async (req, res) => {
             }
         }
 
-        if (successfulInserted === diet.length) {
+        if (successfulInserted === dietArr.length) {
             res.json({ message: 'User diet updated successfully' });
         } else {
             res.status(500).json({ message: 'Error updating user diet' });
@@ -81,21 +115,104 @@ router.post('/:id/diet', async (req, res) => {
     }
 });
 
-router.delete('/:id', (req, res) => {
+router.delete('/:id/diet', (req, res) => {
     const userId = req.params.id;
+    const diets = req.body.diets;
 
-    const sql = 'DELETE FROM user WHERE userId = ?';
-    db.query(sql, [userId])
-        .then((result) => {
-            if (result.affectedRows === 0) {
-                res.status(404).json({ message: 'User not found' });
+    if (!diets) {
+        return res.status(400).json({ message: 'Please provide diets' });
+    }
+
+    const dietArr = diets.split(',').map(diet => diet.trim());
+
+    // Check if the user exists
+    const checkUserSQL = 'SELECT 1 FROM user WHERE userId = ?';
+    db.query(checkUserSQL, [userId])
+        .then((userResult) => {
+            if (userResult.length === 0) {
+                res.status(400).json({ message: 'User not found' });
+                return Promise.reject('User not found');
+            }
+
+            // Check if the user favorites exist
+            const checkUserDietSQL = 'SELECT * FROM userdiet WHERE userId = ? AND diet = ?';
+            const checkPromises = dietArr.map(diet => {
+                return db.query(checkUserDietSQL, [userId, diet]);
+            });
+
+            return Promise.all(checkPromises);
+        })
+        .then((userDietResults) => {
+            const invalidDiets = userDietResults.filter(result => result.length === 0);
+
+            if (invalidDiets.length > 0) {
+                res.status(400).json({ message: 'Diet(s) not found' });
+                return Promise.reject('Diet(s) not found');
+            }
+
+            // Delete user's favorite recipes
+            const deleteSQL = 'DELETE FROM userdiet WHERE userId = ? AND diet IN (?)';
+            return db.query(deleteSQL, [userId, dietArr]);
+        })
+        .then((deleteResult) => {
+            if (deleteResult.affectedRows === dietArr.length) {
+                res.json({ message: 'User diets deleted successfully' });
             } else {
-                res.json({ message: 'User deleted successfully' });
+                res.status(500).json({ message: 'Error deleting user diets' });
             }
         })
         .catch((err) => {
-            res.status(500).json({ error: err.message });
+            if (!res.headersSent) {
+                res.status(500).json({ message: err });
+            }
         });
+});
+
+router.get('/:id/recipe', (req, res) => {
+    const userId = req.params.id;
+
+    // Check if the user exists
+    const checkUserSQL = 'SELECT 1 FROM user WHERE userId = ?';
+    db.query(checkUserSQL, [userId])
+        .then((userResult) => {
+            if (userResult.length === 0) {
+                res.status(400).json({ message: 'User not found' });
+                return;
+            }
+
+            let sql = `
+                SELECT recipe.*, 
+                    (SELECT JSON_ARRAYAGG(name) FROM dishtypes WHERE dishtypes.recipeId = recipe.recipeId) AS dishtypes,
+                    (SELECT JSON_ARRAYAGG(name) FROM diets WHERE diets.recipeId = recipe.recipeId) AS diets,
+                    (SELECT JSON_ARRAYAGG(name) FROM cuisines WHERE cuisines.recipeId = recipe.recipeId) AS cuisines FROM recipe
+                INNER JOIN userfavourite ON recipe.recipeId=userfavourite.recipeId
+                INNER JOIN user ON userfavourite.userId = user.userId
+                WHERE user.userId=?;
+            `;
+
+            db.query(sql, [userId])
+                .then((rows) => {
+                    if (rows.length === 0) {
+                        res.status(404).json({ message: 'No favourite recipes found' });
+                        return;
+                    }
+
+                    rows.forEach((row) => {
+                        row.dishtypes = row.dishtypes;
+                        row.diets = row.diets;
+                        row.cuisines = row.cuisines;
+
+                        row.recipeId = Number(row.recipeId);
+                        row.id = Number(row.id);
+                    });
+
+                    res.json(rows);
+                })
+                .catch((err) => {
+                    console.log(err)
+                    res.status(500).json({ error: err.message });
+                });
+        })
 });
 
 router.post('/:userId/recipe/:recipeId', (req, res) => {
@@ -107,7 +224,7 @@ router.post('/:userId/recipe/:recipeId', (req, res) => {
     db.query(checkUserSQL, [userId])
         .then((userResult) => {
             if (userResult.length === 0) {
-                res.status(404).json({ message: 'User not found' });
+                res.status(400).json({ message: 'User not found' });
                 return Promise.reject('User not found');
             }
 
@@ -117,7 +234,7 @@ router.post('/:userId/recipe/:recipeId', (req, res) => {
         })
         .then((recipeResult) => {
             if (recipeResult.length === 0) {
-                res.status(404).json({ message: 'Recipe not found' });
+                res.status(400).json({ message: 'Recipe not found' });
                 return Promise.reject('Recipe not found');
             }
 
@@ -149,41 +266,28 @@ router.post('/:userId/recipe/:recipeId', (req, res) => {
         });
 });
 
-router.delete('/:userId/recipe/:recipeIds', (req, res) => {
-    const userId = req.params.userId;
+router.delete('/:id/recipe', (req, res) => {
+    const userId = req.params.id;
     const recipeIds = req.body.recipeIds;
-    const recipeIdsArray = recipeIds.split(',').map(d => d.trim()); 
 
-    const sql = `DELETE FROM userfavourite WHERE userId = ? AND recipeId IN (${recipeIdsArray.map(id => `'${id}'`).join(', ')})`;
-    db.query(sql, [userId, recipeId])
-        .then((result) => {
-            if (result.affectedRows === 0) {
-                res.status(500).json({ message: 'Error adding user favourite recipe' });
-            } else {
-                res.json({ message: 'User favourite recipe added successfully' });
-            }
-        })
-        .catch((err) => {
-            res.status(500).json({ error: err.message });
-        });
-});
+    if (!recipeIds) {
+        return res.status(400).json({ message: 'Please provide recipeIds' });
+    }
 
-router.delete('/:userId/recipe', (req, res) => {
-    const userId = req.params.userId;
-    const recipeIds = req.body.recipeIds.split(',').map(id => id.trim());
+    const recipeIdsArr = recipeIds.split(',').map(id => id.trim());
 
     // Check if the user exists
     const checkUserSQL = 'SELECT 1 FROM user WHERE userId = ?';
     db.query(checkUserSQL, [userId])
         .then((userResult) => {
             if (userResult.length === 0) {
-                res.status(404).json({ message: 'User not found' });
+                res.status(400).json({ message: 'User not found' });
                 return Promise.reject('User not found');
             }
 
             // Check if the user favorites exist
             const checkFavoritesSQL = 'SELECT 1 FROM userfavourite WHERE userId = ? AND recipeId = ?';
-            const checkPromises = recipeIds.map(recipeId => {
+            const checkPromises = recipeIdsArr.map(recipeId => {
                 return db.query(checkFavoritesSQL, [userId, recipeId]);
             });
 
@@ -193,13 +297,13 @@ router.delete('/:userId/recipe', (req, res) => {
             const invalidRecipes = favoritesResults.filter(result => result.length === 0);
 
             if (invalidRecipes.length > 0) {
-                res.status(404).json({ message: 'Invalid recipe(s) in user favorites' });
-                return Promise.reject('Invalid recipe(s) in user favorites');
+                res.status(400).json({ message: 'Recipe(s) is not user favorites' });
+                return Promise.reject('Recipe(s) is not user favorites');
             }
 
             // Check if each recipeId is valid
             const checkRecipeSQL = 'SELECT 1 FROM recipe WHERE recipeId = ?';
-            const checkRecipePromises = recipeIds.map(recipeId => {
+            const checkRecipePromises = recipeIdsArr.map(recipeId => {
                 return db.query(checkRecipeSQL, [recipeId]);
             });
 
@@ -209,16 +313,16 @@ router.delete('/:userId/recipe', (req, res) => {
             const invalidRecipes = recipeResults.filter(result => result.length === 0);
 
             if (invalidRecipes.length > 0) {
-                res.status(404).json({ message: 'Invalid recipe(s) in recipe table' });
-                return Promise.reject('Invalid recipe(s) in recipe table');
+                res.status(400).json({ message: 'Recipe(s) not found' });
+                return Promise.reject('Recipe(s) not found');
             }
 
             // Delete user's favorite recipes
             const deleteSQL = 'DELETE FROM userfavourite WHERE userId = ? AND recipeId IN (?)';
-            return db.query(deleteSQL, [userId, recipeIds]);
+            return db.query(deleteSQL, [userId, recipeIdsArr]);
         })
         .then((deleteResult) => {
-            if (deleteResult.affectedRows === recipeIds.length) {
+            if (deleteResult.affectedRows === recipeIdsArr.length) {
                 res.json({ message: 'User favorite recipes deleted successfully' });
             } else {
                 res.status(500).json({ message: 'Error deleting user favorite recipes' });
